hypervisor: refactor VMM overhead into shared library and integrate into KubeVirt
Extract VMM overhead calculation functions from kvm.go into a new shared
file vmm_overhead.go (no build tag) so they are available to all hypervisor
backends:
- vmmOverhead: top-level function respecting global config, per-app override,
  and automatic estimation priority
- estimatedVMMOverhead: combines all overhead components
- ramVMMOverhead: 2.5% of domain RAM for page table overhead
- qemuVMMOverhead: 20 MB for QEMU binaries/libraries
- mmioVMMOverhead: 1% of total MMIO size for GPU/VGA passthrough devices
- cpuVMMOverhead: 3 MB per vCPU
- undefinedVMMOverhead: 350 MB base overhead for QEMU internal use

kvm.go: remove the above functions (now in vmm_overhead.go); CountMemOverhead
wrapper is unchanged.

kubevirt.go: add CountMemOverhead method on kubevirtContext that delegates to
the shared vmmOverhead function, replacing the inherited zero-overhead
ctrdContext implementation. Add uuid import.

include VMM overhead in DomainMetric.AllocatedMB

Previously AllocatedMB in KubeVirt DomainMetric was set to the Kubernetes
resource limit (guest RAM only), so reported metrics and device-level
AllocatedAppsMB undercounted total host memory consumption by the full
VMM overhead.

Calculate and store the VMM overhead (via the shared vmmOverhead function)
in vmiMetaData when the domain config is created — both for VMI replicasets
(CreateReplicaVMIConfig) and pod replicasets (CreateReplicaPodConfig).

In GetDomsCPUMem (VMI virt-handler metrics path), add the stored overhead
to AllocatedMB in the post-processing loop after all metrics are filled.

In getPodMetrics (pod replicaset path), add the stored overhead to
AllocatedMB before constructing the DomainMetric.

This makes KubeVirt consistent with KVM, where AllocatedMB is read from
the cgroup HierarchicalMemoryLimit which naturally includes QEMU overhead.

- updatestatus.go: skip CountMemOverhead call when VirtualizationMode is NOHYPER
- kubevirt.go: skip vmmOverhead in CreateReplicaVMIConfig and
  CreateReplicaPodConfig when VirtualizationMode is NOHYPER

Store globalConfig on kubevirtContext in Setup() so that CreateReplicaVMIConfig
and CreateReplicaPodConfig can pass it to vmmOverhead instead of nil.

This ensures the global VmmMemoryLimitInMiB override is respected when
calculating VMM overhead during domain config creation, consistent with
the CountMemOverhead path.

Replace truncating integer division (memOverhead / (1024 * 1024)) with
roundFromBytesToMbytes() when adding VMM overhead to AllocatedMB. This
avoids under-reporting overhead by up to ~1 MB and is consistent with
how other memory conversions are done in the codebase.
