+++ pkg/pillar/cmd/zedmanager/updatestatus.go
@@ -669,18 +669,24 @@ func doActivate(ctx *zedmanagerContext, uuidStr string,
 
 		// If we have not yet calculated memory overhead - do it now
 		if status.MemOverhead == 0 {
-			// Get hypervisor
-			hyp, err := hypervisor.GetHypervisor(*ctx.hypervisorPtr)
-			if err != nil {
-				log.Fatalf("Cannot get hypervisor: %s", err)
-			}
+			// NOHYPER runs containers directly without a VMM, so there is no hypervisor overhead.
+			if config.FixedResources.VirtualizationMode == types.NOHYPER {
+				log.Functionf("doActivate(%s): skipping VMM overhead for NOHYPER virtualization mode",
+					status.Key())
+			} else {
+				// Get hypervisor
+				hyp, err := hypervisor.GetHypervisor(*ctx.hypervisorPtr)
+				if err != nil {
+					log.Fatalf("Cannot get hypervisor: %s", err)
+				}
 
-			status.MemOverhead, err = hyp.CountMemOverhead(status.DomainName, config.UUIDandVersion.UUID,
-				int64(config.FixedResources.Memory), int64(config.FixedResources.VMMMaxMem),
-				int64(config.FixedResources.MaxCpus), int64(config.FixedResources.VCpus), config.IoAdapterList,
-				ctx.assignableAdapters, ctx.globalConfig)
-			// We have to publish the status here, because we need to save the memory overhead value, it's used in getRemainingMemory
-			publishAppInstanceStatus(ctx, status)
+				status.MemOverhead, err = hyp.CountMemOverhead(status.DomainName, config.UUIDandVersion.UUID,
+					int64(config.FixedResources.Memory), int64(config.FixedResources.VMMMaxMem),
+					int64(config.FixedResources.MaxCpus), int64(config.FixedResources.VCpus), config.IoAdapterList,
+					ctx.assignableAdapters, ctx.globalConfig)
+				// We have to publish the status here, because we need to save the memory overhead value, it's used in getRemainingMemory
+				publishAppInstanceStatus(ctx, status)
+			}
 		}
 
 		remaining, latent, halting, err := getRemainingMemory(ctx)
